requirements.txt
beautifulsoup4~=4.12.3
pillow~=10.4.0
pycryptodome
requests~=2.32.3
requests-oauthlib
requests_kerberos
requests_ntlm
oauthlib~=3.2.2
cryptography~=43.0.0